+++ c/common/include/euler_utils.h
@@ -6,7 +6,7 @@
 
 #define count_representation_combinations(v,m,s) count_representation_combinations_ (v,m,s,NULL)
 
-int euclid_pythagorean_triple_perim (int, int);
+long euclid_pythagorean_triple_perim (int, int);
 long stern_brocot_count (long, const fraction_t *, const fraction_t *);
 int count_representation_combinations_ (const int *, size_t, int, int **);
 char * calculate_sqrt (int, size_t);
